Implement PDF comparison with rudimentary byte-difference display. If a reference file exists, it will be compared with its associated example file. If the reference file is missing, the associated example file is considered to be without differences. PDF files in any subdirectory named reference are not deleted when the test begins. This commit is broken -- it illustrates how the use of maps for fonts, images and other resources lead to different dictionary tables in the PDF document.

+++ fpdf_test.go
@@ -40,9 +40,14 @@ func init() {
 func cleanup() {
 	filepath.Walk(example.PdfDir(),
 		func(path string, info os.FileInfo, err error) (reterr error) {
-			if len(path) > 3 {
-				if path[len(path)-4:] == ".pdf" {
-					os.Remove(path)
+			if info.Mode().IsRegular() {
+				dir, _ := filepath.Split(path)
+				if "reference" != filepath.Base(dir) {
+					if len(path) > 3 {
+						if path[len(path)-4:] == ".pdf" {
+							os.Remove(path)
+						}
+					}
 				}
 			}
 			return
